handler/requirements.txt
Django
psycopg2-binary

python-dateutil
pytz

uWSGI

simpleeval

djangorestframework
djangorestframework-gis
django-filter
django-crispy-forms
markdown
Pygments

celery
django-celery-beat
redis
django-redis
